forecasting/requirements.txt
pandas
numpy
statsmodels
sqlalchemy
pymysql
python-dotenv
